Cast from int64_t to long for printf.
On 32-bit ARM platforms this is required otherwise 0 is printed.

+++ pipe_lat.c
@@ -66,7 +66,7 @@ int main(int argc, char *argv[]) {
   }
 
   printf("message size: %i octets\n", size);
-  printf("roundtrip count: %li\n", count);
+  printf("roundtrip count: %li\n", (long)count);
 
   if (pipe(ofds) == -1) {
     perror("pipe");
@@ -138,7 +138,7 @@ int main(int argc, char *argv[]) {
 
 #endif
 
-    printf("average latency: %li ns\n", delta / (count * 2));
+    printf("average latency: %li ns\n", (long)(delta / (count * 2)));
   }
 
   return 0;
